backend/requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.1
Django==4.2.2
django-cors-headers==4.1.0
django-filter==23.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.6
inflection==0.5.1
Markdown==3.4.3
packaging==23.1
Pillow==10.0.0
pycparser==2.21
pytube==15.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
